requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.0.2
django-cors-headers==4.3.1
django-ninja==1.1.0
gunicorn==21.2.0
mysqlclient==2.2.4
packaging==24.2
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
requests==2.31.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
